EdgeKey retreival code should be with the other EdgeKey code
